Fix NoClassDefFoundError but now VerifyError. we need -noverify
